Work out pin count of spare throw so the frame sum is 10
The new test case knocks down pins in the first throw, so 10 can't be
assumed for the spare throw but for the whole frame.

This makes the parsing trickier as when a spare is seen, the value of
the first throw is needed. However, as the first throw can't be a spare,
split the parsing into 2, so the one for the second throw can have the
first throw as param.