Work out strike looking ahead the next 2 throws
They can come from the next 2 frames if they were strikes, so increase
the window size but flatten them out into throws so that the strike
can be calculated without knowledge of which frames the throws come from
as that detail is not relevant.

Naming the functions that parse throws is a bit tricky, so decided on
making clear which one can end up in a spare.